app/requirements.txt
python-telegram-bot==21.6
python-dotenv
aiohttp
requests
beautifulsoup4
playwright
pandas
lxml
